requirements.txt
tensorflow==2.0.0
opencv-python
Flask==1.1.1
gunicorn
Jinja2==2.10.1
numpy
scipy
scikit-learn
matplotlib
pandas
Pillow
h5py
keras
